Improve parallelization and preview (CONFIG.H updated)
Instead of spawning a thread for each horizontal line of the image, now
only N-1 threads are spawned, where N is the number of threads your
computer can use.
Each thread writes separately to the result.bmp file
and this should (!) not be a problem as they're not writing to the same
areas.
The debug console now shows the amount of seconds remaining.

It is possible to increase the amount of threads to N (instead of N-1),
but beware that this is very demanding for your machine.